pyproject.toml
[project]
name = "scraping-to-excel"
version = "0.1.0"
description = ""
authors = [
    {name = "vini",email = "viniciusdafonseca2000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "patchright (>=1.52.5,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "black (>=25.1.0,<26.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "httpx[http2] (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
